+++ CHANGELOG.md
@@ -5,3 +5,14 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 - Initial release
+
+## [0.1.0] - 2019-02-23
+### Added
+- Install can search by package version.
+### Changed
+- Update README.
+- Reflector code.
+
+## [0.0.1] - 2019-02-19
+### Added
+- Add `init`, `install`, `uninstall`, `restore`, `clean` command